Move InetAddressPredicates and related classes to common package
Motivation:
Their usage are not limited to `server`, so it would be better to be in `common` package.

Modifications:
- Move the classes in `server.misc` to `common.util`.
- Remove `server.misc` package.
