+++ core/src/main/java/com/linecorp/armeria/common/util/Inet4AddressBlock.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.linecorp.armeria.server.misc;
+package com.linecorp.armeria.common.util;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static java.util.Objects.requireNonNull;
